+++ wifi.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define MAX_ESSID_LENGTH 50
 
 typedef struct {
-  unsigned int signal_strength;
+  uint_fast8_t signal_strength;
   char name[MAX_ESSID_LENGTH];
 } WifiInfo;
